[activity-list] Rename CategoryDetailsViewModel by ActivityViewModel & CategoryDetailView by ActivitiesListView - Remove unnecessary code
