+++ TrendyTravel.xcodeproj/xcuserdata/ijoe.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -14,8 +14,8 @@
             filePath = "TrendyTravel/ViewModel/CategoryDetailsViewModel.swift"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "31"
-            endingLineNumber = "31"
+            startingLineNumber = "26"
+            endingLineNumber = "26"
             landmarkName = "getActivities(from:)"
             landmarkType = "7">
          </BreakpointContent>
