+++ TrendyTravel/ViewModel/ActivityViewModel.swift
@@ -1,5 +1,5 @@
 //
-//  CategoryDetailsViewModel.swift
+//  ActivityViewModel.swift
 //  TrendyTravel
 //
 //  Created by Jonathan Duong on 17/07/2023.
@@ -8,11 +8,9 @@
 import Foundation
 import NetworkManager
 
-class CategoryDetailsViewModel: ObservableObject {
-    @Published var isLoading = true
+class ActivityViewModel: ObservableObject {
     @Published var activities: [Activity] = []
     @Published var category: Category = .culture
-    @Published var errorMessage = ""
     
     let urlString = "https://trendytravel.onrender.com/activities"
 
@@ -24,15 +22,10 @@ class CategoryDetailsViewModel: ObservableObject {
                 self.activities = fetchedActivities.filter {
                     $0.category == category.rawValue
                 }
-                if !self.activities.isEmpty {
-                    self.isLoading = false
-                }
                 self.category = category
             }
         } catch {
-            DispatchQueue.main.async {
-                self.errorMessage = error.localizedDescription
-            }
+            print(error)
         }
     }
 }
